when generating target_link_options, don't join flags that span multiple arguments into one string
