+++ tests/files/test_parse_and_generate/clang_install_name/build.log
@@ -1,0 +1,1 @@
+clang -c ../source/foo.c -o foo.oclang -dynamiclib -Wl,-headerpad_max_install_names -Wl,-search_paths_first -install_name /usr/local/lib/libfoo.dylib -o libfoo.dylib foo.o
