+++ tests/files/test_parse_and_generate/path_alias_rpath_gcc/CMakeLists.txt
@@ -9,4 +9,4 @@ set(SOURCE_DIR "${CMAKE_CURRENT_LIST_DIR}" CACHE PATH "")
 set(RPATH_PATH_ROOT /some/path CACHE STRING "")
 
 add_library(a SHARED ${SOURCE_DIR}/main.cpp)
-target_link_options(a PRIVATE "-Wl,-rpath -Wl,${RPATH_PATH_ROOT}/lib")
+target_link_options(a PRIVATE -Wl,-rpath -Wl,${RPATH_PATH_ROOT}/lib)
