+++ tests/files/test_parse_and_generate/shared_from_static_without_sources/CMakeLists.txt
@@ -15,5 +15,5 @@ if(NOT EXISTS ${CMAKE_CURRENT_BINARY_DIR}/empty.c)
     file(WRITE ${CMAKE_CURRENT_BINARY_DIR}/empty.c "")
 endif()
 add_library(impl SHARED ${CMAKE_CURRENT_BINARY_DIR}/empty.c)
-target_link_options(impl PRIVATE "-z noexecstack" -Wl,-z,defs)
+target_link_options(impl PRIVATE -z noexecstack -Wl,-z,defs)
 target_link_libraries(impl PRIVATE m -Wl,-whole-archive impl1.static -Wl,-no-whole-archive)
