+++ tests/files/test_parse_and_generate/shared_from_static_without_sources_for_several_empty_targets/CMakeLists.txt
@@ -15,14 +15,14 @@ if(NOT EXISTS ${CMAKE_CURRENT_BINARY_DIR}/empty.c)
     file(WRITE ${CMAKE_CURRENT_BINARY_DIR}/empty.c "")
 endif()
 add_library(impl_c SHARED ${CMAKE_CURRENT_BINARY_DIR}/empty.c)
-target_link_options(impl_c PRIVATE "-z noexecstack" -Wl,-z,defs)
+target_link_options(impl_c PRIVATE -z noexecstack -Wl,-z,defs)
 target_link_libraries(impl_c PRIVATE m -Wl,-whole-archive impl_c_s.static -Wl,-no-whole-archive)
 
 add_library(impl_cpp_s.static STATIC ${SOURCE_DIR}/impl_cpp.cpp)
 set_target_properties(impl_cpp_s.static PROPERTIES OUTPUT_NAME impl_cpp_s)
 
 add_library(impl_cpp SHARED ${CMAKE_CURRENT_BINARY_DIR}/empty.c)
-target_link_options(impl_cpp PRIVATE "-z noexecstack" -Wl,-z,defs)
+target_link_options(impl_cpp PRIVATE -z noexecstack -Wl,-z,defs)
 target_link_libraries(impl_cpp PRIVATE
     m
     -Wl,-whole-archive
